split 3 more large files: SettingsWindow, ShellIntegration, MainWindow.UI
- SettingsWindow.xaml.cs (1051->602): assoc logic -> SettingsWindow.Assoc.cs (448)
- ShellIntegration.cs (869->99): split into Menu.cs (402) + Assoc.cs (313), shared decls remain
- MainWindow.UI.cs (1111->979): FolderNode/ArchiveItem/CompressedDisplayMode -> MainWindow.Types.cs (139)
- 0 build errors, 171/171 tests pass
