+++ docs/PROGRESS.md
@@ -25,6 +25,9 @@
 9. **修复 .GetAwaiter().GetResult() 同步-异步反模式**：`ResolveSmartDest` 改为 async，用 `await` 替代阻塞
 10. **App.Cli.cs 拆分**：按职责拆为 App.Compress.cs（压缩命令）、App.Extract.cs（解压命令）、App.Open.cs（打开/快速压缩），原文件保留为空白 partial 壳
 11. **CompressSettingsWindow 拆分**：密码标签页逻辑独立为 CompressSettingsWindow.Password.cs partial 文件，主文件减少 450 行
+12. **SettingsWindow.xaml.cs 拆分**：文件关联面板逻辑独立为 SettingsWindow.Assoc.cs partial 文件，主文件从 1051 行降至 602 行
+13. **ShellIntegration.cs 拆分**：拆为 ShellIntegration.Menu.cs（右键菜单注册）+ ShellIntegration.Assoc.cs（文件关联注册），原文件保留共享声明（99 行）
+14. **MainWindow.UI.cs 类型抽取**：FolderNode、ArchiveItem 子类、CompressedDisplayMode 枚举移到 MainWindow.Types.cs（139 行）
 
 1. **文件关联 Bug 修复**：
    - `.tar.gz` 不再被跳过——设置勾选后真正写入注册表 `OpenWithProgids` + `DefaultIcon`
