+++ src/MantisZip.UI/ShellIntegration.cs
@@ -20,7 +20,7 @@ namespace MantisZip.UI;
 ///   6. 压缩到（父目录名）
 ///   7. 用MantisZip压缩
 /// </summary>
-internal static class ShellIntegration
+internal static partial class ShellIntegration
 {
     [DllImport("shell32.dll", CharSet = CharSet.Auto)]
     private static extern void SHChangeNotify(uint wEventId, uint uFlags, IntPtr dwItem1, IntPtr dwItem2);
@@ -68,679 +68,12 @@ internal static class ShellIntegration
     private static readonly string CompressCombinedDisplay = L.T(L.Shell_CompressCombined);
     private static readonly string CompressDisplay = L.T(L.Shell_Compress);
 
-    /// <summary>检查是否已L.T(L.Settings_Menu_Btn_Install) Shell 扩展（COM 或静态注册表）。</summary>
-    public static bool IsInstalled
-    {
-        get
-        {
-            // 检查 COM 注册（优先）
-            using var comKey = Registry.CurrentUser.OpenSubKey(
-                $@"Software\Classes\CLSID\{ComClsid}");
-            if (comKey != null) return true;
-
-            // 回退：检查静态注册表
-            using var key = Registry.CurrentUser.OpenSubKey(
-                $@"Software\Classes\*\shell\{CascadeRoot}");
-            if (key != null) return true;
-            using var key2 = Registry.CurrentUser.OpenSubKey(
-                $@"Software\Classes\*\shell\{CompressVerb}\command");
-            if (key2 != null) return true;
-            using var key3 = Registry.CurrentUser.OpenSubKey(
-                $@"Software\Classes\*\shell\{ExtractSmartVerb}\command");
-            if (key3 != null) return true;
-            using var key4 = Registry.CurrentUser.OpenSubKey(
-                $@"Software\Classes\*\shell\{CompressSeparateVerb}\command");
-            return key4 != null;
-        }
-    }
-
-    /// <summary>
-    /// L.T(L.Settings_Menu_Btn_Install) Shell 右键菜单。
-    /// 使用 COM IContextMenu 实现（MantisZip.ShellExt.comhost.dll），
-    /// 回退到静态注册表方案（旧版兼容）。
-    /// </summary>
-    public static void Install()
-    {
-        App.LogDebug("ShellIntegration.Install: starting");
-        var s = AppSettings.Instance;
-        var exePath = GetExePath();
-
-        // 先清理旧注册，避免残留
-        Uninstall();
-
-        // 尝试安装 COM 组件（MantisZip.ShellExt.comhost.dll）
-        if (InstallCom())
-        {
-            App.LogDebug("ShellIntegration.Install: COM context menu installed");
-        }
-        else
-        {
-            // 回退到静态注册表方案（仅级联模式，动词模式已在 v0.4.0 移除）
-            App.LogDebug("ShellIntegration.Install: COM not available, falling back to static cascade");
-            InstallCascade(s, exePath);
-        }
-
-        // 通知 Windows Shell 刷新上下文菜单缓存
-        SHChangeNotify(SHCNE_ASSOCCHANGED, SHCNF_IDLIST, IntPtr.Zero, IntPtr.Zero);
-        App.LogDebug("ShellIntegration.Install: done, exePath={0}", exePath);
-    }
-
-    /// <summary>
-    /// L.T(L.Settings_Menu_Btn_Uninstall)所有 L.T(L.App_MantisZipTitle) 右键菜单注册。
-    /// </summary>
-    public static void Uninstall()
-    {
-        App.LogDebug("ShellIntegration.Uninstall: starting");
-        // 层叠入口
-        foreach (var target in new[] { "*", "Directory", @"Directory\Background" })
-        {
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{CascadeRoot}");
-        }
-
-        // 层叠子命令定义
-        DeleteRegistryKey($@"Software\Classes\{CascadeRoot}");
-
-        // 独立动词（含新旧 verb L.T(L.Main_Col_Name)全L.T(L.CompressConflict_Overwrite)，升级时清理旧版注册）
-        foreach (var target in new[] { "*", "Directory", @"Directory\Background" })
-        {
-            // 新名称
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{OpenVerb}");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{ExtractHereVerb}");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{ExtractSmartVerb}");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{ExtractToNamedVerb}");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{ExtractVerb}");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{CompressSeparateVerb}");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{CompressCombinedVerb}");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\{CompressVerb}");
-
-            // 菜单分隔线（独立动词模式）
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\00_MantisZipSepTop");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\02a_MantisZipSep");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\04a_MantisZipSep");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\99_MantisZipSepBottom");
-
-            // 旧版动词L.T(L.Main_Col_Name)（v0.1.3 及以前）
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\MantisZipOpen");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\MantisZipExtract");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\MantisZipQuick");
-            DeleteRegistryKey($@"Software\Classes\{target}\shell\MantisZipCompress");
-        }
-
-        // 清理 COM 组件注册
-        UninstallCom();
-
-        // 旧版 per-extension 注册（v0.1.3 早期版本遗留）
-        foreach (var ext in ArchiveExtensions)
-        {
-            DeleteRegistryKey($@"Software\Classes\{ext}\shell\MantisZipExtract");
-        }
-        // 通知 Windows Shell 刷新上下文菜单缓存
-        SHChangeNotify(SHCNE_ASSOCCHANGED, SHCNF_IDLIST, IntPtr.Zero, IntPtr.Zero);
-
-        App.LogDebug("ShellIntegration.Uninstall: done");
-    }
-
-    #region COM 组件注册/反注册
-
     // MantisZip.ShellExt 组件 CLSID（必须与 ContextMenuHandler.cs 的 [Guid] 一致）
     private const string ComClsid = "{C90B2A1E-5E4F-4A7A-9B0F-8C1D3E5F7A9B}";
     private const string ComProgId = "MantisZip.ContextMenu";
     private const string ComHandlerKey = "MantisZip";
 
-    /// <summary>
-    /// 安装 COM 右键菜单（MantisZip.ShellExt.comhost.dll）。
-    /// 写入 HKCU\Software\Classes，无需管理员权限。
-    /// 返回 true 表示安装成功，false 表示 COM host DLL 不存在（触发回退）。
-    /// </summary>
-    private static bool InstallCom()
-    {
-        try
-        {
-            var baseDir = Path.GetDirectoryName(GetExePath());
-            if (baseDir == null) return false;
-
-            var comhostPath = Path.Combine(baseDir, "MantisZip.ShellExt.comhost.dll");
-            if (!File.Exists(comhostPath))
-            {
-                // Also check subdirectory (x64/x86 for architecture-specific builds)
-                comhostPath = Path.Combine(baseDir,
-                    Environment.Is64BitProcess ? "x64" : "x86",
-                    "MantisZip.ShellExt.comhost.dll");
-                if (!File.Exists(comhostPath))
-                {
-                    App.LogDebug("InstallCom: comhost.dll not found at {0}", comhostPath);
-                    return false;
-                }
-            }
-
-            App.LogDebug("InstallCom: registering from {0}", comhostPath);
-
-            // 1. CLSID 注册（HKCU — per-user）
-            //    .NET 6+ comhost.dll 通过 runtimeconfig.json 定位运行时和托管程序集，
-            //    不需要 InprocServer32 下额外的 Assembly/Class/CodeBase 值
-            var clsidKey = $@"Software\Classes\CLSID\{ComClsid}";
-            SetRegistryValue(clsidKey, null, "MantisZip Context Menu Handler");
-            SetRegistryValue($@"{clsidKey}\InprocServer32", null, comhostPath);
-            SetRegistryValue($@"{clsidKey}\InprocServer32", "ThreadingModel", "Apartment");
-            SetRegistryValue($@"{clsidKey}\ProgId", null, ComProgId);
-            // ProgId 反向查找（由 COM CreateObject/ProgId 需要）
-            SetRegistryValue($@"Software\Classes\{ComProgId}", null, "MantisZip Context Menu Handler");
-            SetRegistryValue($@"Software\Classes\{ComProgId}\CLSID", null, ComClsid);
-
-            // 2. 上下文菜单处理程序注册
-            foreach (var target in new[] { "*", "Directory", @"Directory\Background" })
-            {
-                var handlerKey = $@"Software\Classes\{target}\shellex\ContextMenuHandlers\{ComHandlerKey}";
-                SetRegistryValue(handlerKey, null, ComClsid);
-            }
-
-            App.LogDebug("InstallCom: COM registration successful");
-            WriteMenuTextToRegistry();
-            return true;
-        }
-        catch (Exception ex)
-        {
-            App.LogDebug("InstallCom: failed: {0}", ex.Message);
-            return false;
-        }
-    }
-
-    /// <summary>
-    /// 卸载 COM 右键菜单，并清理旧的静态注册表条目。
-    /// </summary>
-    private static void UninstallCom()
-    {
-        try
-        {
-            // 1. 删除上下文菜单处理程序注册
-            foreach (var target in new[] { "*", "Directory", @"Directory\Background" })
-            {
-                DeleteRegistryKey($@"Software\Classes\{target}\shellex\ContextMenuHandlers\{ComHandlerKey}");
-            }
-
-            // 2. 删除 CLSID 注册及 ProgId 反向查找
-            DeleteRegistryKey($@"Software\Classes\CLSID\{ComClsid}");
-            DeleteRegistryKey($@"Software\Classes\{ComProgId}");
-
-            App.LogDebug("UninstallCom: COM registration cleaned up");
-        }
-        catch (Exception ex)
-        {
-            App.LogDebug("UninstallCom: failed: {0}", ex.Message);
-        }
-    }
-
-    #endregion
-
-    #region 层叠子菜单模式 (ExtendedSubCommandsKey)
-
-    private static void InstallCascade(AppSettings s, string exePath)
-    {
-        // 每个目标类使用独立的子命令 key，因为 %1 / %V 不同
-        InstallCascadeFor("*",                  "File",      s, exePath, argVar: "%1", includeExtract: true);
-        InstallCascadeFor("Directory",           "Directory", s, exePath, argVar: "%1", includeExtract: false);
-        InstallCascadeFor(@"Directory\Background", "Background", s, exePath, argVar: "%V", includeExtract: false);
-    }
-
-    private static void InstallCascadeFor(string target, string subKeySuffix, AppSettings s, string exePath, string argVar, bool includeExtract)
-    {
-        // 顶级层叠入口：MUIVerb + ExtendedSubCommandsKey
-        var entryPath = $@"Software\Classes\{target}\shell\{CascadeRoot}";
-        SetRegistryValue(entryPath, "MUIVerb", L.T(L.App_MantisZipTitle));
-
-        var subCommandsPath = $@"{CascadeRoot}\{subKeySuffix}";
-        SetRegistryValue(entryPath, "ExtendedSubCommandsKey", subCommandsPath);
-
-        if (s.ShowMenuIcons)
-            SetRegistryValue(entryPath, "Icon", $"""{exePath},0""");
-
-        // 子命令：{CascadeRoot}\{suffix}\shell\{order_name}\command
-        // 注：分隔线使用独立的 separator verb（CommandFlags=8），
-        // 不直接设置在真实动词上，避免部分 Windows 版本上
-        // ECF_SEPARATORBEFORE 导致动词本身不显示。
-        var shellPath = $@"Software\Classes\{subCommandsPath}\shell";
-        int order = 0;
-
-        // 1. 用MantisZip打开（所有文件均可，由应用层处理非压缩包）
-        if (s.EnableOpenMenu)
-        {
-            order++;
-            var verb = $"{order:D2}_open";
-            var verbPath = $@"{shellPath}\{verb}";
-            SetRegistryValue(verbPath, null, OpenDisplay);
-            if (s.ShowMenuIcons)
-            {
-                string? icon = GetMenuIconPath(verb);
-                if (icon != null) SetRegistryValue(verbPath, "Icon", icon);
-            }
-            SetRegistryValue($@"{verbPath}\command", null, $@"""{exePath}"" --open ""{argVar}""");
-        }
-
-        // 解压相关动词（所有文件均可，跟压缩一样）
-        if (includeExtract)
-        {
-            bool hasExtract = s.EnableExtractHereMenu || s.EnableSmartExtractMenu || s.EnableExtractToNamedMenu || s.EnableExtractToMenu;
-            if (hasExtract)
-            {
-                order++;
-                InstallSeparator(shellPath, order);
-            }
-
-            // 2. 用MantisZip解压到此处
-            if (s.EnableExtractHereMenu)
-            {
-                order++;
-                var verb = $"{order:D2}_extracthere";
-                var verbPath = $@"{shellPath}\{verb}";
-                SetRegistryValue(verbPath, null, ExtractHereDisplay);
-                if (s.ShowMenuIcons)
-                {
-                    string? icon = GetMenuIconPath(verb);
-                    if (icon != null) SetRegistryValue(verbPath, "Icon", icon);
-                }
-                SetRegistryValue($@"{verbPath}\command", null, $@"""{exePath}"" --extract-here ""{argVar}""");
-            }
-
-            // 2.5. 智能解压到此处 (Smart Extract)
-            if (s.EnableSmartExtractMenu)
-            {
-                order++;
-                var verb = $"{order:D2}_smartextract";
-                var verbPath = $@"{shellPath}\{verb}";
-                SetRegistryValue(verbPath, null, ExtractSmartDisplay);
-                if (s.ShowMenuIcons)
-                {
-                    string? icon = GetMenuIconPath(verb);
-                    if (icon != null) SetRegistryValue(verbPath, "Icon", icon);
-                }
-                SetRegistryValue($@"{verbPath}\command", null, $@"""{exePath}"" --extract-smart ""{argVar}""");
-            }
-
-            // 3. 用MantisZip解压到（压缩包名）
-            if (s.EnableExtractToNamedMenu)
-            {
-                order++;
-                var verb = $"{order:D2}_extracttonamed";
-                var verbPath = $@"{shellPath}\{verb}";
-                SetRegistryValue(verbPath, null, ExtractToNamedDisplay);
-                if (s.ShowMenuIcons)
-                {
-                    string? icon = GetMenuIconPath(verb);
-                    if (icon != null) SetRegistryValue(verbPath, "Icon", icon);
-                }
-                SetRegistryValue($@"{verbPath}\command", null, $@"""{exePath}"" --extract-to-name ""{argVar}""");
-            }
-
-            // 4. 用MantisZip解压到……
-            if (s.EnableExtractToMenu)
-            {
-                order++;
-                var verb = $"{order:D2}_extract";
-                var verbPath = $@"{shellPath}\{verb}";
-                SetRegistryValue(verbPath, null, ExtractDisplay);
-                if (s.ShowMenuIcons)
-                {
-                    string? icon = GetMenuIconPath(verb);
-                    if (icon != null) SetRegistryValue(verbPath, "Icon", icon);
-                }
-                SetRegistryValue($@"{verbPath}\command", null, $@"""{exePath}"" --extract ""{argVar}""");
-            }
-        }
-
-        // 压缩
-        {
-            bool hasCompress = s.EnableCompressSeparate || s.EnableCompressCombined || s.EnableCompressMenu;
-            if (hasCompress)
-            {
-                order++;
-                InstallSeparator(shellPath, order);
-            }
-
-            // 5. 压缩到独立的（文件名）
-            if (s.EnableCompressSeparate)
-            {
-                order++;
-                var verb = $"{order:D2}_compressseparate";
-                var verbPath = $@"{shellPath}\{verb}";
-                SetRegistryValue(verbPath, null, CompressSeparateDisplay);
-                if (s.ShowMenuIcons)
-                {
-                    string? icon = GetMenuIconPath(verb);
-                    if (icon != null) SetRegistryValue(verbPath, "Icon", icon);
-                }
-                SetRegistryValue($@"{verbPath}\command", null, $@"""{exePath}"" --compress-separate ""{argVar}""");
-            }
-
-            // 6. 压缩到（父目录名）
-            if (s.EnableCompressCombined)
-            {
-                order++;
-                var verb = $"{order:D2}_compresscombined";
-                var verbPath = $@"{shellPath}\{verb}";
-                SetRegistryValue(verbPath, null, CompressCombinedDisplay);
-                if (s.ShowMenuIcons)
-                {
-                    string? icon = GetMenuIconPath(verb);
-                    if (icon != null) SetRegistryValue(verbPath, "Icon", icon);
-                }
-                SetRegistryValue($@"{verbPath}\command", null, $@"""{exePath}"" --compress-combined ""{argVar}""");
-            }
-
-            // 7. 用MantisZip压缩
-            if (s.EnableCompressMenu)
-            {
-                order++;
-                var verb = $"{order:D2}_compress";
-                var verbPath = $@"{shellPath}\{verb}";
-                SetRegistryValue(verbPath, null, CompressDisplay);
-                if (s.ShowMenuIcons)
-                {
-                    string? icon = GetMenuIconPath(verb);
-                    if (icon != null) SetRegistryValue(verbPath, "Icon", icon);
-                }
-                SetRegistryValue($@"{verbPath}\command", null, $@"""{exePath}"" --compress ""{argVar}""");
-            }
-        }
-    }
-
-    #endregion
-
-    #region L.T(L.Settings_Tab_FileAssoc)
-
-    /// <summary>检查L.T(L.Settings_Tab_FileAssoc)L.T(L.MsgBox_Yes)L.T(L.MsgBox_No)已L.T(L.Settings_Menu_Btn_Install)。</summary>
-    public static bool AreAssociationsInstalled => ArchiveExtensions.Any(AreAssociationsInstalledForExtension);
-
-    /// <summary>检查指定的扩展名是否已安装文件关联。</summary>
-    public static bool AreAssociationsInstalledForExtension(string ext)
-    {
-        var progId = GetProgId(ext);
-        using var key = Registry.CurrentUser.OpenSubKey($@"Software\Classes\{ext}\OpenWithProgids");
-        if (key == null) return false;
-        return key.GetValue(progId) != null;
-    }
-
-    /// <summary>检查指定扩展名的关联状态（无关联/关联未默认/关联且默认）。</summary>
-    public static AssocStatus GetAssociationStatus(string ext)
-    {
-        if (!AreAssociationsInstalledForExtension(ext))
-            return AssocStatus.NotAssociated;
-
-        try
-        {
-            using var key = Registry.CurrentUser.OpenSubKey($@"Software\Microsoft\Windows\CurrentVersion\Explorer\FileExts\{ext}\UserChoice");
-            var progId = key?.GetValue("Progid") as string;
-            if (!string.IsNullOrEmpty(progId) && progId.IndexOf("MantisZip", StringComparison.OrdinalIgnoreCase) >= 0)
-                return AssocStatus.IsDefault;
-        }
-        catch { }
-
-        return AssocStatus.NotDefault;
-    }
-
-    /// <summary>为单个扩展名安装文件关联（OpenWithProgids + DefaultIcon + 独立 ProgId）。</summary>
-    public static void InstallAssociationForExtension(string ext)
-    {
-        var progId = GetProgId(ext);
-
-        // 1. 注册独立 ProgId（含具体图标和打开命令）
-        EnsureProgIdRegistered(ext);
-
-        // 2. 写 OpenWithProgids
-        var extKey = $@"Software\Classes\{ext}\OpenWithProgids";
-        using var key = Registry.CurrentUser.CreateSubKey(extKey);
-        key?.SetValue(progId, Array.Empty<byte>(), RegistryValueKind.None);
-
-        // 3. 写 DefaultIcon
-        var iconPath = GetIconPath(ext);
-        if (iconPath != null)
-        {
-            SetRegistryValue($@"Software\Classes\{ext}\DefaultIcon", null, iconPath);
-        }
-        else
-        {
-            // 自定义扩展名 — 使用软件自身图标
-            var exePath = GetExePath();
-            if (exePath != null)
-                SetRegistryValue($@"Software\Classes\{ext}\DefaultIcon", null, $@"""{exePath}"",0");
-        }
-    }
-
-    /// <summary>注册 Applications 条目（全局，仅一次）。</summary>
-    private static void EnsureApplicationsRegistered()
-    {
-        App.LogDebug("ShellIntegration.EnsureApplicationsRegistered: starting");
-        var exePath = GetExePath();
-
-        var appKey = $@"Software\Classes\Applications\{Path.GetFileName(exePath)}";
-        SetRegistryValue(appKey, "FriendlyAppName", L.T(L.App_MantisZipTitle));
-        SetRegistryValue($@"{appKey}\shell\open\command", null, $@"""{exePath}"" --open ""%1""");
-        foreach (var ext in ArchiveExtensions)
-        {
-            if (ext == ".tar.gz") continue;
-            SetRegistryValue($@"{appKey}\SupportedTypes", ext, "");
-        }
-
-        App.LogDebug("ShellIntegration.EnsureApplicationsRegistered: done");
-    }
-
-    /// <summary>注册一个扩展名的独立 ProgId（含图标和打开命令）。</summary>
-    private static void EnsureProgIdRegistered(string ext)
-    {
-        var progId = GetProgId(ext);
-        var exePath = GetExePath();
-        var progIdKey = $@"Software\Classes\{progId}";
-
-        // 只有需要创建时才写入，避免覆盖用户的自定义设置
-        using var check = Registry.CurrentUser.OpenSubKey(progIdKey);
-        if (check != null) return;
-
-        // 友好名称
-        var displayName = L.T(L.Shell_ProgIdDesc) + " — " + ext.TrimStart('.');
-        SetRegistryValue(progIdKey, null, displayName);
-        SetRegistryValue($@"{progIdKey}\shell\open", null, L.T(L.Shell_OpenVerb));
-        SetRegistryValue($@"{progIdKey}\shell\open\command", null, $@"""{exePath}"" --open ""%1""");
-
-        // 具体格式图标
-        var iconPath = GetIconPath(ext);
-        if (iconPath != null)
-            SetRegistryValue($@"{progIdKey}\DefaultIcon", null, iconPath);
-        else
-            SetRegistryValue($@"{progIdKey}\DefaultIcon", null, $@"""{exePath},0""");
-    }
-
-    /// <summary>清理旧的 MantisZip.Archive 单 ProgId（v0.5 以前版本迁移用）。</summary>
-    public static void CleanupOldProgId()
-    {
-        DeleteRegistryKey($@"Software\Classes\{OldProgId}");
-    }
-
-    /// <summary>安装文件关联前的预备工作：迁移旧版 + 注册 Applications 条目。</summary>
-    public static void PrepareAssocRegistration()
-    {
-        CleanupOldProgId();
-        EnsureApplicationsRegistered();
-    }
-
-    /// <summary>
-    /// L.T(L.Shell_Compress)格式扩展名逐个关联独立 ProgId。
-    /// 写入 HKCU\Software\Classes，无需管理员权限。
-    /// </summary>
-    public static void InstallAssociations()
-    {
-        App.LogDebug("ShellIntegration.InstallAssociations: starting");
-
-        // 清理旧的单 ProgId（v0.5 以前版本迁移）
-        CleanupOldProgId();
-
-        // 注册 Applications 条目
-        EnsureApplicationsRegistered();
-
-        // 逐个安装独立 ProgId + 文件关联
-        foreach (var ext in ArchiveExtensions)
-        {
-            InstallAssociationForExtension(ext);
-        }
-
-        App.LogDebug("ShellIntegration.InstallAssociations: done");
-    }
-
-    /// <summary>删除单个扩展名的文件关联（OpenWithProgids + DefaultIcon + 独立 ProgId）。</summary>
-    public static void UninstallAssociationForExtension(string ext)
-    {
-        var progId = GetProgId(ext);
-
-        // 删除 OpenWithProgids 条目
-        try
-        {
-            using var key = Registry.CurrentUser.OpenSubKey($@"Software\Classes\{ext}\OpenWithProgids", writable: true);
-            key?.DeleteValue(progId, throwOnMissingValue: false);
-        }
-        catch { /* 该扩展可能没有 OpenWithProgids 键 */ }
-
-        // 删除 DefaultIcon（仅当指向我们的图标路径时）
-        try
-        {
-            var iconPath = GetIconPath(ext);
-            var exePath = GetExePath();
-            using var extKey = Registry.CurrentUser.OpenSubKey($@"Software\Classes\{ext}", writable: true);
-            if (extKey?.GetValue(null) is string currentIcon)
-            {
-                if (iconPath != null && currentIcon.Equals(iconPath, StringComparison.OrdinalIgnoreCase))
-                {
-                    extKey.DeleteValue(null!);
-                }
-                else if (exePath != null && currentIcon.Equals($@"""{exePath}"",0", StringComparison.OrdinalIgnoreCase))
-                {
-                    extKey.DeleteValue(null!);
-                }
-            }
-        }
-        catch { }
-
-        // 删除独立 ProgId 键（仅当该 ProgId 不被其他扩展引用时）
-        DeleteRegistryKey($@"Software\Classes\{progId}");
-    }
-
-    /// <summary>
-    /// L.T(L.Settings_Assoc_Uninstall)：删除所有独立 ProgId 和扩展名的 OpenWithProgids 条目。
-    /// </summary>
-    public static void UninstallAssociations()
-    {
-        App.LogDebug("ShellIntegration.UninstallAssociations: starting");
-
-        // 删除每个内置扩展名的 OpenWithProgids + DefaultIcon + 独立 ProgId
-        foreach (var ext in ArchiveExtensions)
-        {
-            UninstallAssociationForExtension(ext);
-        }
-
-        // 也清理自定义扩展名的注册表条目
-        try
-        {
-            foreach (var ext in AppSettings.Instance.CustomAssocExtensions)
-            {
-                UninstallAssociationForExtension(ext);
-            }
-        }
-        catch { }
-
-        // 删除 Applications 条目
-        var exeName = Path.GetFileName(Environment.ProcessPath ?? "");
-        if (!string.IsNullOrEmpty(exeName))
-            DeleteRegistryKey($@"Software\Classes\Applications\{exeName}");
-
-        // 清理旧版单 ProgId（v0.5 以前）
-        DeleteRegistryKey($@"Software\Classes\{OldProgId}");
-
-        App.LogDebug("ShellIntegration.UninstallAssociations: done");
-    }
-
-    /// <summary>返回已安装文件关联的扩展名数量（仅统计 UI 中可见的扩展名）。</summary>
-    public static int GetInstalledExtensionCount()
-    {
-        // .tgz 不在 BuiltinAssocFormats 中，不计入 UI 状态总数
-        return ArchiveExtensions
-            .Where(e => e != ".tgz")
-            .Count(AreAssociationsInstalledForExtension);
-    }
-
-    /// <summary>读取指定扩展名的当前默认打开程序，返回友好显示名称。</summary>
-    public static string GetCurrentHandler(string ext)
-    {
-        try
-        {
-            var userChoice = $@"Software\Microsoft\Windows\CurrentVersion\Explorer\FileExts\{ext}\UserChoice";
-            using var key = Registry.CurrentUser.OpenSubKey(userChoice);
-            var progId = key?.GetValue("Progid") as string;
-
-            if (string.IsNullOrEmpty(progId))
-                return "未设置";
-
-            // 检查是否是我们
-            if (progId.Contains("MantisZip", StringComparison.OrdinalIgnoreCase))
-                return "MantisZip";
-
-            // Applications\xxx.exe 格式
-            if (progId.StartsWith("Applications\\", StringComparison.OrdinalIgnoreCase))
-            {
-                var exeName = progId.Substring("Applications\\".Length);
-                return Path.GetFileNameWithoutExtension(exeName);
-            }
-
-            // 尝试从 ProgId 获取友好名称
-            try
-            {
-                using var progIdKey = Registry.ClassesRoot.OpenSubKey(progId);
-                var displayName = progIdKey?.GetValue(null) as string;
-                if (!string.IsNullOrEmpty(displayName))
-                    return CleanAppName(displayName);
-            }
-            catch { }
-
-            // 去除扩展名后缀："Bandizip.zip" → "Bandizip"
-            var knownExts = new[] { ".zip", ".7z", ".rar", ".tar", ".gz", ".tgz", ".iso" };
-            foreach (var knownExt in knownExts)
-            {
-                if (progId.EndsWith(knownExt, StringComparison.OrdinalIgnoreCase))
-                    return progId.Substring(0, progId.Length - knownExt.Length);
-            }
-
-            return progId;
-        }
-        catch
-        {
-            return "未设置";
-        }
-    }
-
-    /// <summary>清理 ProgId 友好名称中常见的格式描述后缀。</summary>
-    private static string CleanAppName(string name)
-    {
-        if (string.IsNullOrEmpty(name)) return name;
-
-        // Strip common suffixes that make the result look like a format description
-        string[] suffixes = { " Archive", " Compressed", " File", " Document", " Image" };
-        foreach (var suffix in suffixes)
-        {
-            if (name.EndsWith(suffix, StringComparison.OrdinalIgnoreCase))
-                return name.Substring(0, name.Length - suffix.Length);
-        }
-
-        return name;
-    }
-
-    #endregion
-
-    #region 辅助方法
-
-    private static string BuildAppliesToFilter()
-    {
-        // .tar.gz 文件的扩展名是 .gz，由 .gz 条目覆盖
-        var exts = ArchiveExtensions
-            .Where(e => e != ".tar.gz")
-            .Select(e => $"System.FileExtension:=\"{e}\"");
-        return string.Join(" OR ", exts);
-    }
+    #region Registry helpers
 
     private static void SetRegistryValue(string subKey, string? valueName, string value)
     {
@@ -754,34 +87,9 @@ internal static class ShellIntegration
         key?.SetValue(valueName, value, RegistryValueKind.DWord);
     }
 
-    /// <summary>在指定位置添加一条菜单分隔线。</summary>
-    private static void InstallSeparator(string shellPath, int order)
-    {
-        var verb = $"{order:D2}_sep";
-        SetRegistryDword($@"{shellPath}\{verb}", "CommandFlags", 8);
-    }
-
     private static void DeleteRegistryKey(string subKey)
     {
         Registry.CurrentUser.DeleteSubKeyTree(subKey, throwOnMissingSubKey: false);
-    }
-
-    /// <summary>
-    /// Write localized ShellExt menu text to registry under HKCU\Software\MantisZip\ContextMenu.
-    /// Called during COM install so ShellExt reads the current language's strings.
-    /// </summary>
-    private static void WriteMenuTextToRegistry()
-    {
-        var regPath = @"Software\MantisZip\ContextMenu";
-        SetRegistryValue(regPath, "TextOpen", L.T(L.ShellExt_Open));
-        SetRegistryValue(regPath, "TextExtractHereSingle", L.T(L.ShellExt_ExtractHereSingle));
-        SetRegistryValue(regPath, "TextExtractHereMulti", L.T(L.ShellExt_ExtractHereMulti));
-        SetRegistryValue(regPath, "TextSmartExtractSingle", L.T(L.ShellExt_SmartExtractSingle));
-        SetRegistryValue(regPath, "TextSmartExtractMulti", L.T(L.ShellExt_SmartExtractMulti));
-        SetRegistryValue(regPath, "TextExtractToNamed", L.T(L.ShellExt_ExtractToNamed));
-        SetRegistryValue(regPath, "TextExtractTo", L.T(L.ShellExt_ExtractTo));
-        SetRegistryValue(regPath, "TextCompress", L.T(L.ShellExt_Compress));
-        App.LogDebug("WriteMenuTextToRegistry: wrote ShellExt menu text to registry");
     }
 
     private static string GetExePath()
@@ -798,72 +106,5 @@ internal static class ShellIntegration
         return assemblyPath;
     }
 
-    /// <summary>
-    /// 返回扩展名对应的图标文件绝对路径。
-    /// 图标文件位于输出目录的 Resources\Icons\ 下。
-    /// 如果图标文件不存在，返回 null。
-    /// </summary>
-    private static string? GetIconPath(string extension)
-    {
-        var iconName = extension.ToLowerInvariant() switch
-        {
-            ".zip" => "zip.ico",
-            ".7z" => "sevenz.ico",
-            ".rar" => "rar.ico",
-            ".tar" => "tar.ico",
-            ".tgz" or ".tar.gz" => "tgz.ico",
-            ".gz" => "gz.ico",
-            ".iso" => "iso.ico",
-            _ => null,
-        };
-        if (iconName == null) return null;
-
-        var baseDir = Path.GetDirectoryName(GetExePath());
-        if (baseDir == null) return null;
-
-        var iconPath = Path.Combine(baseDir, "Resources", "Icons", iconName);
-        return File.Exists(iconPath) ? iconPath : null;
-    }
-
-    /// <summary>
-    /// 返回菜单动词对应的图标文件绝对路径（与 COM 动态菜单一致）。
-    /// 动词名形如 "01_open"、"02_extracthere"，自动忽略前缀匹配。
-    /// 图标文件位于输出目录的 Resources\MenuIcons\ 下。
-    /// 如果图标文件不存在，返回 null。
-    /// </summary>
-    private static string? GetMenuIconPath(string verb)
-    {
-        // Strip order prefix ("01_open" → "open")
-        int underscoreIdx = verb.IndexOf('_');
-        string key = underscoreIdx >= 0 ? verb.Substring(underscoreIdx + 1) : verb;
-
-        var iconName = key.ToLowerInvariant() switch
-        {
-            "open" => "Open.ico",
-            "extracthere" => "ExtractHere.ico",
-            "smartextract" => "ExtractSmart.ico",
-            "extracttonamed" => "ExtractToNamed.ico",
-            "extract" => "ExtractTo.ico",
-            "compressseparate" => "CompressSeparate.ico",
-            "compresscombined" => "CompressCombined.ico",
-            "compress" => "CompressDialog.ico",
-            _ => null,
-        };
-        if (iconName == null) return null;
-
-        var baseDir = Path.GetDirectoryName(GetExePath());
-        if (baseDir == null) return null;
-
-        var iconPath = Path.Combine(baseDir, "Resources", "MenuIcons", iconName);
-        return File.Exists(iconPath) ? iconPath : null;
-    }
-
     #endregion
-}
-
-public enum AssocStatus
-{
-    NotAssociated,
-    NotDefault,
-    IsDefault
 }
